pyproject.toml
[tool.poetry]
name = "caiso_connector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "2.22.0"
pytest = "5.3.2"
pandas = "0.25.3"
numpy = "1.18.0"
fastapi = "^0.45.0"
uvicorn = "^0.11.1"
aiohttp = "^3.6"
aiofiles = "^0.4.0"
python-crontab = "^2.4"

[tool.poetry.dev-dependencies]
jupyter = "^1.0"
altair = "^4.0"
vega_datasets = "^0.8.0"
vega = "=1.3"
sphinx = "^2.3"
sphinx_rtd_theme = "^0.4.3"
recommonmark = "^0.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
